Better Multi-Version Handling
+ Add list of full SCLs to PDK families
+ Added ability to specify which SCLs to push
~ Moved management-related CLI functions to `__main__.py`
~ Fixed github token requirement check
~ `manage.py::enable` no longer attempts to exit on its own, raises Exceptions like a responsible citizen
- Removed `os.EX_` exits- not available on Windows
