+++ volare/__init__.py
@@ -11,8 +11,8 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
+from .manage import VersionNotFound, enable
 from .common import get_volare_home, get_current_version, get_installed_list, Version
-from .manage import enable
 from .build import build
 
 from .__version__ import __version__
