+++ volare/manage.py
@@ -11,17 +11,15 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
+import io
 import os
-import sys
-import json
 import shutil
 import tarfile
 import requests
 import tempfile
-from typing import Any, List, Optional
+from typing import Any, List, Optional, Union
 
 import rich
-import click
 import rich.tree
 import rich.progress
 import zstandard as zstd
@@ -30,20 +28,17 @@ from rich.console import Console
 from .build.git_multi_clone import mkdirp
 from .common import (
     Version,
-    check_version,
     get_release_links,
     get_versions_dir,
     get_volare_dir,
     get_installed_list,
 )
-from .click_common import (
-    opt,
-    opt_build,
-    opt_push,
-    opt_pdk_root,
-)
 from .build import build, push
 from .families import Family
+
+
+class VersionNotFound(Exception):
+    pass
 
 
 def print_installed_list(
@@ -110,128 +105,6 @@ def print_remote_list(
     console.print(tree)
 
 
-# -- CLI
-@click.command("output")
-@opt_pdk_root
-def output_cmd(pdk_root, pdk):
-    """Outputs the currently enabled PDK version.
-
-    If not outputting to a tty, the output is either the version string
-    unembellished, or, if no current version is enabled, an empty output with an
-    exit code of 1.
-    """
-
-    version = Version.get_current(pdk_root, pdk)
-    if sys.stdout.isatty():
-        if version is None:
-            print(f"No version of the PDK {pdk} is currently enabled at {pdk_root}.")
-            print(
-                "Invoke volare --help for assistance installing and enabling versions."
-            )
-            exit(1)
-        else:
-            print(f"Installed: {pdk} v{version.name}")
-            print(
-                "Invoke volare --help for assistance installing and enabling versions."
-            )
-    else:
-        if version is None:
-            exit(1)
-        else:
-            print(version.name, end="")
-
-
-@click.command("prune")
-@opt_pdk_root
-@click.option(
-    "--yes",
-    is_flag=True,
-    callback=lambda c, _, v: not v and c.abort(),
-    expose_value=False,
-    prompt="Are you sure? This will delete all non-enabled versions of the PDK from your computer.",
-)
-def prune_cmd(pdk_root, pdk):
-    """Removes all PDKs other than, if it exists, the one currently in use."""
-    pdk_versions = get_installed_list(pdk_root, pdk)
-    for version in pdk_versions:
-        if version.is_current(pdk_root):
-            continue
-        try:
-            version.uninstall()
-            print(f"Deleted {version}.")
-        except Exception as e:
-            print(f"Failed to delete {version}: {e}", file=sys.stderr)
-
-
-@click.command("rm")
-@opt_pdk_root
-@click.option(
-    "--yes",
-    is_flag=True,
-    callback=lambda c, _, v: not v and c.abort(),
-    expose_value=False,
-    prompt="Are you sure? This will delete this version of the PDK from your computer.",
-)
-@click.argument("version", required=False)
-def rm_cmd(pdk_root, pdk, version):
-    """Removes the PDK version specified."""
-    version_object = Version(version, pdk)
-    try:
-        version_object.uninstall(pdk_root)
-        print(f"Deleted {version}.")
-    except Exception as e:
-        print(f"Failed to delete: {e}", file=sys.stderr)
-        exit(1)
-
-
-@click.command("ls")
-@opt_pdk_root
-def list_cmd(pdk_root, pdk):
-    """Lists PDK versions that are locally installed. JSON if not outputting to a tty."""
-
-    pdk_versions = get_installed_list(pdk_root, pdk)
-
-    if sys.stdout.isatty():
-        console = Console()
-        print_installed_list(pdk_root, pdk, console, pdk_versions)
-    else:
-        print(json.dumps([version.name for version in pdk_versions]), end="")
-
-
-@click.command("ls-remote")
-@opt_pdk_root
-def list_remote_cmd(pdk_root, pdk):
-    """Lists PDK versions that are remotely available. JSON if not outputting to a tty."""
-
-    try:
-        all_versions = Version._from_github()
-        pdk_versions = all_versions.get(pdk) or []
-
-        if sys.stdout.isatty():
-            console = Console()
-            print_remote_list(pdk_root, pdk, console, pdk_versions)
-        else:
-            print(json.dumps([version.name for version in pdk_versions]), end="")
-    except requests.exceptions.ConnectionError:
-        if sys.stdout.isatty():
-            console = Console()
-            console.print(
-                "[red]You don't appear to be connected to the Internet. ls-remote cannot be used."
-            )
-        else:
-            print("Failed to connect to remote server", file=sys.stderr)
-        sys.exit(-1)
-
-
-@click.command("path")
-@opt_pdk_root
-@click.argument("version", required=False)
-def path_cmd(pdk_root, pdk, version):
-    """Prints the path of a specific pdk version installation."""
-    version = Version(version, pdk)
-    print(version.get_dir(pdk_root), end="")
-
-
 def enable(
     pdk_root: str,
     pdk: str,
@@ -241,9 +114,11 @@ def enable(
     build_kwargs: dict = {},
     push_kwargs: dict = {},
     include_libraries: Optional[List[str]] = None,
+    output: Union[Console, io.TextIOWrapper] = Console(),
 ):
-    console = Console()
-
+    console = output
+    if not isinstance(console, Console):
+        console = Console(file=console)
     current_file = os.path.join(get_volare_dir(pdk_root, pdk), "current")
     current_file_dir = os.path.dirname(current_file)
     mkdirp(current_file_dir)
@@ -254,8 +129,7 @@ def enable(
 
     pdk_family = Family.by_name.get(pdk)
     if pdk_family is None:
-        print(f"Unsupported PDK family '{pdk}'.", file=sys.stderr)
-        exit(os.EX_USAGE)
+        raise ValueError(f"Unsupported PDK family '{pdk}'.")
 
     variants = pdk_family.variants
 
@@ -272,10 +146,9 @@ def enable(
                 if also_push:
                     push(pdk_root=pdk_root, pdk=pdk, version=version, **push_kwargs)
             else:
-                console.print(
-                    f"[red]Version {version} not found either locally or remotely.\nTry volare build {version}."
+                raise FileNotFoundError(
+                    f"Version {version} not found either locally or remotely."
                 )
-                exit(1)
         else:
             try:
                 tarball_directory = tempfile.TemporaryDirectory(suffix=".volare")
@@ -315,19 +188,18 @@ def enable(
                                 mkdirp(final_dir)
                                 io = tf.extractfile(file)
                                 if io is None:
-                                    raise ValueError(
-                                        f"Failed to unpack tarball for {name}."
+                                    raise IOError(
+                                        f"Failed to unpack file in {name}'s tarball: {file.name}."
                                     )
                                 with open(final_path, "wb") as f:
                                     f.write(io.read())
-            except Exception as e:
-                shutil.rmtree(version_directory, ignore_errors=True)
-                console.print(f"[red]Error: {e}")
-                exit(-1)
-            except KeyboardInterrupt:
+            except KeyboardInterrupt as e:
                 console.print("Interrupted.")
                 shutil.rmtree(version_directory, ignore_errors=True)
-                exit(-1)
+                raise e from None
+            except Exception as e:
+                shutil.rmtree(version_directory, ignore_errors=True)
+                raise e from None
 
             for variant in variants:
                 variant_install_path = os.path.join(version_directory, variant)
@@ -344,10 +216,9 @@ def enable(
                 if os.path.islink(path):
                     os.unlink(path)
                 else:
-                    console.print(
-                        f"[red]Error: {path} exists, and not as a symlink. Please manually remove it before continuing."
+                    raise FileExistsError(
+                        f"{path} exists, and not as a symlink. Remove it first."
                     )
-                    exit(1)
 
         for vpath, fpath in zip(version_paths, final_paths):
             src = os.path.relpath(vpath, pdk_root)
@@ -357,106 +228,3 @@ def enable(
             f.write(version)
 
     console.print(f"Version {version} enabled for the {pdk} PDK.")
-
-
-@click.command("enable")
-@opt_pdk_root
-@click.option(
-    "-f",
-    "--metadata-file",
-    "tool_metadata_file_path",
-    default=None,
-    help="Explicitly define a tool metadata file instead of searching for a metadata file",
-)
-@click.option(
-    "-l",
-    "--include-libraries",
-    multiple=True,
-    default=None,
-    help="Libraries to include. You can use -l multiple times to include multiple libraries. Pass 'all' to include all of them. A default of 'None' uses a default set for the particular PDK.",
-)
-@click.argument("version", required=False)
-def enable_cmd(pdk_root, pdk, tool_metadata_file_path, version, include_libraries):
-    """
-    Activates a given installed PDK version.
-
-    Parameters: <version> (Optional)
-
-    If a version is not given, and you run this in the top level directory of
-    tools with a tool_metadata.yml file, for example OpenLane or DFFRAM,
-    the appropriate version will be enabled automatically.
-    """
-    if include_libraries == ():
-        include_libraries = None
-
-    console = Console()
-    version = check_version(version, tool_metadata_file_path, console)
-    enable(
-        pdk_root=pdk_root,
-        pdk=pdk,
-        version=version,
-        include_libraries=include_libraries,
-    )
-
-
-@click.command("enable_or_build", hidden=True)
-@opt_pdk_root
-@opt_push
-@opt_build
-@opt("--also-push/--dont-push", default=False, help="Also push.")
-@click.option(
-    "-f",
-    "--metadata-file",
-    "tool_metadata_file_path",
-    default=None,
-    help="Explicitly define a tool metadata file instead of searching for a metadata file",
-)
-@click.argument("version")
-def enable_or_build_cmd(
-    include_libraries,
-    jobs,
-    pdk_root,
-    pdk,
-    owner,
-    repository,
-    token,
-    pre,
-    clear_build_artifacts,
-    tool_metadata_file_path,
-    also_push,
-    version,
-    use_repo_at,
-    build_magic,
-):
-    """
-    Attempts to activate a given PDK version. If the version is not found locally or remotely,
-    it will instead attempt to build said version.
-
-    Parameters: <version>
-    """
-    if include_libraries == ():
-        include_libraries = None
-
-    console = Console()
-    version = check_version(version, tool_metadata_file_path, console)
-    enable(
-        pdk_root=pdk_root,
-        pdk=pdk,
-        version=version,
-        build_if_not_found=True,
-        also_push=also_push,
-        build_kwargs={
-            "include_libraries": include_libraries,
-            "jobs": jobs,
-            "clear_build_artifacts": clear_build_artifacts,
-            "use_repo_at": use_repo_at,
-            "build_magic": build_magic,
-        },
-        push_kwargs={
-            "owner": owner,
-            "repository": repository,
-            "token": token,
-            "pre": pre,
-        },
-        include_libraries=include_libraries,
-    )
